Remove grand auto bug
Deresute had a bug where, in grand mode, if a slide starts on the rightmost 5 lanes while under a boosted perfect support, when the perfect support drops, the remaining checkpoints are not removed.

This simulator copied that above bug.

Let's remove this feature since the bug has been fixed in deresute.